components/runners/ambient-runner/pyproject.toml
[project]
name = "ambient-runner"
version = "0.4.0"
description = "Polymorphic AG-UI server — Claude Code, Gemini CLI, Codex SDK"
readme = "CLAUDE.md"
requires-python = ">=3.12"
authors = [
  { name = "vTeam" }
]
dependencies = [
  # Ambient Runner SDK core
  "fastapi>=0.135.1",
  "uvicorn[standard]>=0.41.0",
  "ag-ui-protocol>=0.1.13",
  "pydantic>=2.12.5",
  "aiohttp>=3.13.4",
  "requests>=2.32.5",
  "pyjwt>=2.11.0",
  "pyyaml>=6.0.3",
]

[project.optional-dependencies]
claude = [
  "anthropic[vertex]>=0.88.0",
  "claude-agent-sdk>=0.1.50",
]
observability = [
  "langfuse>=3.14.5",
]
mlflow-observability = [
  "mlflow[kubernetes]>=3.11.0",
  "opentelemetry-exporter-otlp>=1.40.0",
]
mcp-atlassian = [
  "mcp-atlassian>=0.21.0",
]
# Install everything for the polymorphic runner
all = [
  "ambient-runner[claude,observability,mlflow-observability,mcp-atlassian]",
]

[dependency-groups]
dev = [
  "pytest>=9.0.2",
  "pytest-asyncio>=1.3.0",
  "pytest-cov>=7.0.0",
  "ruff>=0.15.5",
  "black>=26.3.0",
  "httpx>=0.28.1",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"

[tool.setuptools]
py-modules = ["main"]
packages = ["ag_ui_claude_sdk", "ag_ui_gemini_cli", "ambient_runner", "ambient_runner.endpoints", "ambient_runner.middleware", "ambient_runner.platform", "ambient_runner.bridges", "ambient_runner.bridges.claude", "ambient_runner.bridges.langgraph", "ambient_runner.bridges.gemini_cli"]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"
